- Implemented Leave group functionality.

+++ app/src/main/java/com/example/admin/myapplication/controller/database/remote/GroupMembersDB.java
@@ -125,7 +125,7 @@ public class GroupMembersDB {
                     handler.onObjectReceived(0);
                 }
                 else {
-                    handler.onObjectReceived(dataSnapshot.getChildrenCount());
+                    handler.onObjectReceived(((Long)dataSnapshot.getChildrenCount()).intValue());
                 }
             }
